refactor(api): use centralized buckets constant
