+++ src/config/constants.ts
@@ -1,3 +1,4 @@
+// Shared constants for backend code
 export const SUPABASE_BUCKETS = {
   recipes: 'recipe-images',
   avatars: 'avatars',
